fixed bug that hibernate return backpointer list full of copy
The problem is the fact that when hibernate load a EAGER list, it join
table with other table and when two table has a ciclic reference it
duplicate the column and rename it and after do the join.
Now the problem is that hibernate now look at at the object loaded by
the query and now it see lot of duplicate that has the right id and so
it return all the copy.
For fix the problem:
- add the hashCode() and equals() function to the data class
- change the backpointer list into HashSet
It's all.